+++ Assets/Scripts/EntityComponent/Behavior/Behaviors/Move.cs
@@ -39,7 +39,7 @@ namespace Behavior
                 }
                 else
                 {
-                    Vector3 moved = (Destination - startPosition).normalized * Entity.FactoredMovementSpeed * DeltaTime;
+                    Vector3 moved = (Destination - startPosition).normalized * Entity.FactoredMovementSpeed * (float)DeltaTime;
                     Entity.Position += moved;
                     return true;
                 }
