+++ Assets/Scripts/Game/Game.cs
@@ -108,11 +108,11 @@ namespace LOP
             }
         }
 
-        private void OnUpdateElapsedTime(float time)
+        private void OnUpdateElapsedTime(double time)
         {
             SceneMessageBroker.Publish(new TickMessage.BeforePhysicsSimulation(CurrentTick));
 
-            Physics.Simulate(time);
+            Physics.Simulate((float)time);
 
             SceneMessageBroker.Publish(new TickMessage.AfterPhysicsSimulation(CurrentTick));
         }
